extracts signee list error component to its own file
